Configurator now parses floats, instead of double. Again, this is with regards to BBB performance

+++ CentralComputing/Configurator.cpp
@@ -38,7 +38,7 @@ bool Configurator::getValue(const string& varName, int64_t& value) {
   return true;
 }
 
-bool Configurator::getValue(const string& varName, double& value) {
+bool Configurator::getValue(const string& varName, float& value) {
   if (mapVals.find(varName) == mapVals.end()) {
     return false;
   }
